Move non-dom interacting methods into their own file
