+++ lib/index.js
@@ -1,9 +1,4 @@
-const getTrivia = (m, d) => {
-  fetch(`http://numbersapi.com/${m}/${d}/date?json`)
-  .then(response => response.json())
-  .then(responseData => showTrivia(responseData))
-  .catch(error => console.error({ error }));
-}
+import {randomMonth, randomDay, checkLength, getTrivia} from 'index';
 
 const showTrivia = (trivia) => {
   $('.trivia-container').append(
@@ -12,26 +7,6 @@ const showTrivia = (trivia) => {
     </div>
     `
   )
-}
-
-const randomMonth = () => {
-  let num = Math.floor(Math.random() * 12) + 1
-  let stringedNum = num.toString()
-  return checkLength(stringedNum)
-}
-
-const randomDay = () => {
-  let num = Math.floor(Math.random() * 31) + 1
-  let stringedNum = num.toString()
-  return checkLength(stringedNum)
-}
-
-const checkLength = (stringedNum) => {
-  if(stringedNum.length < 2) {
-    return `0${stringedNum}`
-  } else {
-    return stringedNum
-  }
 }
 
 $('.submit-button').on('click', () => {
